# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hack-Flow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /hackathons/{hackathonId}/teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List teams in a hackathon
      parameters:
        - name: hackathonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a team
      parameters:
        - name: hackathonId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/{id}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get team details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
    patch:
      operationId: updateTeam
      tags: [teams]
      summary: Update team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/{id}/invite:
    post:
      operationId: inviteToTeam
      tags: [teams]
      summary: Invite user to team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInviteInput"
      responses:
        "200":
          description: Invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamInvitation"
  /team-invitations/{id}/respond:
    post:
      operationId: respondToInvitation
      tags: [teams]
      summary: Accept or decline team invitation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvitationResponse"
      responses:
        "200":
          description: Responded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /teams/{id}/leave:
    post:
      operationId: leaveTeam
      tags: [teams]
      summary: Leave a team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Left
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # MATCHMAKING
  /hackathons/{hackathonId}/matchmaking/candidates:
    get:
      operationId: getMatchCandidates
      tags: [matchmaking]
      summary: Get AI-scored team match candidates
      parameters:
        - name: hackathonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match candidates with scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchCandidate"
  /hackathons/{hackathonId}/matchmaking/swipe:
    post:
      operationId: swipeCandidate
      tags: [matchmaking]
      summary: Swipe left or right on a match candidate
      parameters:
        - name: hackathonId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwipeInput"
      responses:
        "200":
          description: Swiped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # PROJECTS
...
  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      summary: List support tickets
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: hackathonId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create a support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
    patch:
      operationId: updateTicket
      tags: [tickets]
      summary: Update ticket status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}/claim:
    post:
      operationId: claimTicket
      tags: [tickets]
      summary: Mentor claims a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}/resolve:
    post:
      operationId: resolveTicket
      tags: [tickets]
      summary: Resolve a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"

  # CHAT
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # AUTH
...
    User:
      type: object
      required: [id, email, name, role, isActive, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [participant, organizer, mentor, jury, admin]
        isActive:
          type: boolean
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        timezone:
          type: ["string", "null"]
        githubUrl:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Team:
      type: object
      required: [id, hackathonId, name, createdAt]
      properties:
        id:
          type: integer
        hackathonId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        lookingForRoles:
          type: array
          items:
            type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/TeamMember"
        leaderId:
          type: integer
        createdAt:
          type: string
          format: date-time
    TeamMember:
      type: object
      required: [userId, role, joinedAt]
      properties:
        userId:
          type: integer
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
        skills:
          type: array
          items:
            type: string
        joinedAt:
          type: string
          format: date-time
    TeamInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        lookingForRoles:
          type: array
          items:
            type: string
    TeamUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        avatarUrl:
          type: string
        lookingForRoles:
          type: array
          items:
            type: string
    TeamInviteInput:
      type: object
      required: [userId]
      properties:
        userId:
          type: integer
        message:
          type: string
    TeamInvitation:
      type: object
      required: [id, teamId, userId, status, createdAt]
      properties:
        id:
          type: integer
        teamId:
          type: integer
        teamName:
          type: string
        userId:
          type: integer
        message:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, accepted, declined]
        createdAt:
          type: string
          format: date-time
    InvitationResponse:
      type: object
      required: [accept]
      properties:
        accept:
          type: boolean

    # MATCHMAKING
    MatchCandidate:
      type: object
      required: [user, compatibilityScore, matchedSkills]
      properties:
        user:
          $ref: "#/components/schemas/User"
        compatibilityScore:
          type: number
          format: float
        matchedSkills:
          type: array
          items:
            type: string
        complementaryRoles:
          type: array
          items:
            type: string
        timezoneCompatibility:
          type: number
          format: float
    SwipeInput:
      type: object
      required: [targetUserId, direction]
      properties:
        targetUserId:
          type: integer
        direction:
          type: string
          enum: [left, right]

    # PROJECTS
...
    Ticket:
      type: object
      required: [id, hackathonId, authorId, title, status, createdAt]
      properties:
        id:
          type: integer
        hackathonId:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        mentorId:
          type: ["integer", "null"]
        mentorName:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, claimed, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high]
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: ["string", "null"]
          format: date-time
    TicketInput:
      type: object
      required: [hackathonId, title, description]
      properties:
        hackathonId:
          type: integer
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
    TicketUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, claimed, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high]

    # CHAT
...