# lib/api-zod/src/generated/types/teamInvitation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hack-Flow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TeamInvitationStatus } from './teamInvitationStatus';

export interface TeamInvitation {
  id: number;
  teamId: number;
  teamName?: string;
  userId: number;
  /** @nullable */
  message?: string | null;
  status: TeamInvitationStatus;
  createdAt: Date;
}