# lib/api-zod/src/generated/types/teamUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hack-Flow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TeamUpdate {
  name?: string;
  description?: string;
  avatarUrl?: string;
  lookingForRoles?: string[];
}